Compute stability of outcome rate over time

Compute stability of outcome rate over time

computeTimeStability(studyPopulation, sccsModel = NULL, alpha = 0.05)

Arguments

studyPopulation
An object created using the createStudyPopulation() function.

sccsModel
Optional: A fitted SCCS model as created using fitSccsModel(). If the model contains splines for seasonality and or calendar time these will be adjusted for before computing stability.

alpha
The alpha (type 1 error) used to test for stability. A Bonferroni correction will be applied for the number of months tested.

Value

A tibble with information on the temporal stability per month. The column stable indicates whether the rate of the outcome is within the expected range for that month, assuming the rate is constant over time.

Details

Computes for each calendar month the rate of the outcome, and evaluates whether that rate is constant over time. If splines are used to adjust for seasonality and/or calendar time, these adjustments are taken into consideration. For each month a two-sided p-value is computed against the null hypothesis that the rate in that month equals the mean rate. This p-value is compared to an alpha value, using a Bonferroni correction to adjust for the multiple testing across months.