createCalendarTimeCovariateSettings {SelfControlledCaseSeries}	R Documentation

Create calendar time settings

Description

Create calendar time settings

Usage

createCalendarTimeCovariateSettings(
  calendarTimeKnots = 5,
  allowRegularization = FALSE,
  computeConfidenceIntervals = FALSE
)

Arguments

calendarTimeKnots
If a single number is provided this is assumed to indicate the number of knots to use for the spline, and the knots are automatically spaced according to equal percentiles of the data. If a series of dates is provided these are assumed to be the exact location of the knots.

allowRegularization
When fitting the model, should the covariates defined here be allowed to be regularized?

computeConfidenceIntervals
Should confidence intervals be computed for the covariates defined here? Setting this to FALSE might save computing time when fitting the model. Will be turned to FALSE automatically when allowRegularization = TRUE.

Details

Create an object specifying whether and how calendar time should be included in the model. Calendar time can be included by splitting patient time into calendar months. During a month, the relative risk attributed to calendar time is assumed to be constant, and the risk from month to month is modeled using a cubic spline.

Whereas the seasonality covariate uses a cyclic spline, repeating every year, this calendar time covariate can model trends over years.

Value

An object of type seasonalitySettings.

[Package SelfControlledCaseSeries version 4.2.0 Index]