createExposure {SelfControlledCaseSeries}	R Documentation

Create exposure definition

Description

Create exposure definition

Usage

createExposure(exposureId, exposureIdRef = "exposureId", trueEffectSize = NA)

Arguments

exposureId
An integer used to identify the exposure in the exposure cohort table.

exposureIdRef
A string used to refer to the exposure when defining covariates using the createEraCovariateSettings() function.

trueEffectSize
For negative and positive controls: the known true effect size. To be used for empirical calibration. Negative controls have trueEffectSize = 1. If the true effect size is unknown, use trueEffectSize = NA.

Details

Create an exposure definition, to be used with the createExposuresOutcome function.

Value

An object of type Exposure.

[Package SelfControlledCaseSeries version 4.2.0 Index]