createExposuresOutcome {SelfControlledCaseSeries}	R Documentation

Create a exposures-outcome combination.

Description

Create a exposures-outcome combination.

Usage

createExposuresOutcome(outcomeId, exposures)

Arguments

outcomeId
An integer used to identify the outcome in the outcome cohort table.

exposures
A list of object of type Exposure as created by createExposure().

Details

Create a set of hypotheses of interest, to be used with the runSccsAnalyses function.

Value

An object of type ExposuresOutcome.

[Package SelfControlledCaseSeries version 4.2.0 Index]