createSccsAnalysis {SelfControlledCaseSeries}	R Documentation

Create a SelfControlledCaseSeries analysis specification

Description

Create a SelfControlledCaseSeries analysis specification

Usage

createSccsAnalysis(
  analysisId = 1,
  description = "",
  getDbSccsDataArgs,
  createStudyPopulationArgs,
  createIntervalDataArgs = NULL,
  fitSccsModelArgs
)

Arguments

analysisId
An integer that will be used later to refer to this specific set of analysis choices.

description
A short description of the analysis.

getDbSccsDataArgs
An object representing the arguments to be used when calling the getDbSccsData function.

createStudyPopulationArgs
An object representing the arguments to be used when calling the getDbSccsData function.

createIntervalDataArgs
An object representing the arguments to be used when calling the createSccsIntervalData or createScriIntervalData function.

fitSccsModelArgs
An object representing the arguments to be used when calling the fitSccsModel function.

Value

An object of type SccsAnalysis, to be used with the runSccsAnalyses function.

[Package SelfControlledCaseSeries version 4.2.0 Index]