createSccsIntervalData {SelfControlledCaseSeries}	R Documentation

Create SCCS era data

Description

Create SCCS era data

Usage

createSccsIntervalData(
  studyPopulation,
  sccsData,
  eraCovariateSettings,
  ageCovariateSettings = NULL,
  seasonalityCovariateSettings = NULL,
  calendarTimeCovariateSettings = NULL,
  minCasesForAgeSeason = NULL,
  minCasesForTimeCovariates = 10000,
  eventDependentObservation = FALSE
)

Arguments

studyPopulation
An object created using the createStudyPopulation() function.

sccsData
An object of type SccsData as created using the getDbSccsData function.

eraCovariateSettings
Either an object of type EraCovariateSettings as created using the createEraCovariateSettings() function, or a list of such objects.

ageCovariateSettings
An object of type ageCovariateSettings as created using the createAgeCovariateSettings() function.

seasonalityCovariateSettings
An object of type seasonalityCovariateSettings as created using the createSeasonalityCovariateSettings() function.

calendarTimeCovariateSettings
An object of type calendarTimeCovariateSettings as created using the createCalendarTimeCovariateSettings() function.

minCasesForAgeSeason
DEPRECATED: Use minCasesForTimeCovariates instead.

minCasesForTimeCovariates
Minimum number of cases to use to fit age, season and calendar time splines. If needed (and available), cases that are not exposed will be included.

eventDependentObservation
Should the extension proposed by Farrington et al. be used to adjust for event-dependent observation time?

Details

This function creates covariates based on the data in the sccsData argument, according to the provided settings. It chops patient time into periods during which all covariates remain constant. The output details these periods, their durations, and a sparse representation of the covariate values.

Value

An object of type SccsIntervalData.

References

Farrington, C. P., Anaya-Izquierdo, A., Whitaker, H. J., Hocine, M.N., Douglas, I., and Smeeth, L. (2011). Self-Controlled case series analysis with event-dependent observation periods. Journal of the American Statistical Association 106 (494), 417-426

[Package SelfControlledCaseSeries version 4.2.0 Index]