cyclicSplineDesign {SelfControlledCaseSeries}	R Documentation

Create a design matrix for a cyclic spline

Description

Create a design matrix for a cyclic spline

Usage

cyclicSplineDesign(x, knots, ord = 4)

Arguments

x
Vector of coordinates of the points to be interpolated.

knots
Location of the knots.

ord
Order of the spline function.

Details

This function is used by other functions in this package.

[Package SelfControlledCaseSeries version 4.2.0 Index]