exportToCsv {SelfControlledCaseSeries}	R Documentation

Export SCCSresults to CSV files

Description

Export SCCSresults to CSV files

Usage

exportToCsv(
  outputFolder,
  exportFolder = file.path(outputFolder, "export"),
  databaseId = 1,
  minCellCount = 5,
  sccsDiagnosticThresholds = createSccsDiagnosticThresholds()
)

Arguments

outputFolder
The folder where runCmAnalyses() generated all results.

exportFolder
The folder where the CSV files will written.

databaseId
A unique ID for the database. This will be appended to most tables.

minCellCount
To preserve privacy: the minimum number of subjects contributing to a count before it can be included in the results. If the count is below this threshold, it will be set to -minCellCount.

sccsDiagnosticThresholds
An object of type SccsDiagnosticThresholds as created using createSccsDiagnosticThresholds().

Details

This requires that runSccsAnalyses() has been executed first. It exports all the results in the outputFolder to CSV files for sharing with other sites.

Value

Does not return anything. Is called for the side-effect of populating the exportFolder with CSV files.

[Package SelfControlledCaseSeries version 4.2.0 Index]