getFileReference {SelfControlledCaseSeries}	R Documentation

Get file reference

Description

Get file reference

Usage

getFileReference(outputFolder)

Arguments

outputFolder
Name of the folder where all the outputs have been written to.

Value

A tibble containing the names of the files for various artifacts created for each outcome-exposures set.

[Package SelfControlledCaseSeries version 4.2.0 Index]