SelfControlledCaseSeries-package {SelfControlledCaseSeries}	R Documentation

SelfControlledCaseSeries: Self-Controlled Case Series

Description

SelfControlledCaseSeries is an R package for performing self- controlled case series (SCCS) analyses in an observational database in the OMOP Common Data Model. It extracts all necessary data from the database and transforms it to the format required for SCCS. Age and season can be modeled using splines assuming constant hazard within calendar months. Event-dependent censoring of the observation period can be corrected for. Many exposures can be included at once (MSCCS), with regularization on all coefficients except for the exposure of interest.

Author(s)

Maintainer: Martijn Schuemie schuemie@ohdsi.org

Authors:

Patrick Ryan

Trevor Shaddox

Marc Suchard

See Also

Useful links:

https://github.com/OHDSI/SelfControlledCaseSeries

Report bugs at https://github.com/OHDSI/SelfControlledCaseSeries/issues

[Package SelfControlledCaseSeries version 4.2.0 Index]