Create era covariate settings

Create era covariate settings

createEraCovariateSettings(
  includeEraIds = NULL,
  excludeEraIds = NULL,
  label = "Covariates",
  stratifyById = TRUE,
  start = 0,
  startAnchor = "era start",
  end = 0,
  endAnchor = "era end",
  firstOccurrenceOnly = FALSE,
  splitPoints = c(),
  allowRegularization = FALSE
)

Arguments

includeEraIds
One or more IDs of variables in the SccsData object that should be used to construct this covariate. If no IDs are specified, all variables will be used.

excludeEraIds
One or more IDs of variables in the [SccsData] object that should not be used to construct this covariate.

label
A label used to identify the covariates created using these settings.

stratifyById
Should a single covariate be created for every ID in the SccsData object, or should a single covariate be constructed? For example, if the IDs identify exposures to different drugs, should a covariate be constructed for every drug, or a single covariate for exposure to any of these drugs. Note that overlap will be considered a single exposure.

start
The start of the risk window (in days) relative to the startAnchor.

startAnchor
The anchor point for the start of the risk window. Can be "era start" or "era end".

end
The end of the risk window (in days) relative to the endAnchor.

endAnchor
The anchor point for the end of the risk window. Can be "era start" or "era end".

firstOccurrenceOnly
Should only the first occurrence of the exposure be used?

splitPoints
To split the risk window into several smaller windows, specify the end of each sub- window relative to the start of the main risk window. If add ExposedDaysToStart is TRUE, the split points will be considered to be relative to the end of the main risk window instead.

allowRegularization
When fitting the model, should the covariates defined here be allowed to be regularized?

Value

An object of type EraCovariateSettings.

Details

Create an object specifying how to create a (set of) era-based covariates.