createEraCovariateSettings {SelfControlledCaseSeries}	R Documentation

Create era covariate settings

Description

Create era covariate settings

Usage

createEraCovariateSettings(
  includeEraIds = NULL,
  excludeEraIds = NULL,
  label = "Covariates",
  stratifyById = FALSE,
  start = 0,
  startAnchor = "era start",
  end = 0,
  endAnchor = "era end",
  firstOccurrenceOnly = FALSE,
  allowRegularization = FALSE,
  profileLikelihood = FALSE,
  exposureOfInterest = FALSE
)

Arguments

includeEraIds
One or more IDs of variables in the SccsData object that should be used to construct this covariate. If no IDs are specified, all variables will be used.

excludeEraIds
One or more IDs of variables in the [SccsData] object that should not be used to construct this covariate.

label
A label used to identify the covariates created using these settings.

stratifyById
Should a single covariate be created for every ID in the SccsData object, or should a single covariate be constructed? For example, if the IDs identify exposures to different drugs, should a covariate be constructed for every drug, or a single covariate for exposure to any of these drugs. Note that overlap will be considered a single exposure.

start
The start of the risk window (in days) relative to the startAnchor.

startAnchor
The anchor point for the start of the risk window. Can be "era start" or "era end".

end
The end of the risk window (in days) relative to the endAnchor.

endAnchor
The anchor point for the end of the risk window. Can be "era start" or "era end".

firstOccurrenceOnly
Should only the first occurrence of the exposure be used?

allowRegularization
When fitting the model, should the covariates defined here be allowed to be regularized?

profileLikelihood
When fitting the model, should the likelihood profile be computed for the covariate defined here? The likelihood profile can be used to avoid making normal approximations on the likelihood and can be used in methods specifically designed to make use of the profile, but may take a while to compute.

exposureOfInterest
If TRUE, the fitted coefficient for this variable will be reported when using runSccsAnalyses(). Requires includeEraIds to be a exposure reference ID as defined in createExposure().

Details

Create an object specifying how to create a (set of) era-based covariates.

Value

An object of type EraCovariateSettings.

[Package SelfControlledCaseSeries version 4.2.0 Index]