createFitSccsModelArgs {SelfControlledCaseSeries}	R Documentation

Create a parameter object for the function fitSccsModel

Description

Create a parameter object for the function fitSccsModel

Usage

createFitSccsModelArgs(
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", selectorType = "byPid", startingVariance =
    0.1, seed = 1, resetCoefficients = TRUE, noiseLevel = "quiet"),
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10))
)

Arguments

prior
The prior used to fit the model. See Cyclops::createPrior for details.

control
The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See Cyclops::createControl for details.

profileGrid
A one-dimensional grid of points on the log(relative risk) scale where the likelihood for coefficient of variables is sampled. See details.

profileBounds
The bounds (on the log relative risk scale) for the adaptive sampling of the likelihood function.

Details

Create an object defining the parameter values.

[Package SelfControlledCaseSeries version 4.2.0 Index]