Create a parameter object for the function getDbSccsData

Create a parameter object for the function getDbSccsData

createGetDbSccsDataArgs(
  useCustomCovariates = FALSE,
  useNestingCohort = FALSE,
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 100,
  studyStartDate = "",
  studyEndDate = "",
  maxCasesPerOutcome = 0,
  exposureIds = "exposureId",
  customCovariateIds = ""
)

Arguments

useCustomCovariates
Create covariates from a custom table?

useNestingCohort
Should the study be nested in a cohort (e.g. people witha specific indication)? If not, the study will be nestedin the general population.

nestingCohortId
A cohort definition ID identifying the records in thenestingCohortTable to use as nesting cohort.

deleteCovariatesSmallCount
The minimum count for a covariate to appear in the data to bekept.

studyStartDate
A calendar date specifying the minimum date where data isused. Date format is 'yyyymmdd'.

studyEndDate
A calendar date specifying the maximum date where data isused. Date format is 'yyyymmdd'.

maxCasesPerOutcome
If there are more than this number of cases for a singleoutcome cases will be sampled to this size. maxCasesPerOutcome = 0indicates no maximum size.

exposureIds
A list of identifiers to define the exposures of interest. IfexposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID.If exposureTable <> DRUG_ERA, exposureIds is used to selectthe cohort_concept_id in the cohort-like table. If noexposureIds are provided, all drugs or cohorts in theexposureTable are included as exposures.

customCovariateIds
A list of cohort definition IDS identifying the records inthe customCovariateTable to use for building customcovariates.

Details

Create an object defining the parameter values.