createGetDbSccsDataArgs {SelfControlledCaseSeries}	R Documentation

Create a parameter object for the function getDbSccsData

Description

Create a parameter object for the function getDbSccsData

Usage

createGetDbSccsDataArgs(
  useCustomCovariates = FALSE,
  useNestingCohort = FALSE,
  nestingCohortId = NULL,
  deleteCovariatesSmallCount = 0,
  studyStartDate = "",
  studyEndDate = "",
  maxCasesPerOutcome = 0,
  exposureIds = "exposureId",
  customCovariateIds = ""
)

Arguments

useCustomCovariates
Create covariates from a custom table?

useNestingCohort
Should the study be nested in a cohort (e.g. people with a specific indication)? If not, the study will be nested in the general population.

nestingCohortId
A cohort definition ID identifying the records in the nestingCohortTable to use as nesting cohort.

deleteCovariatesSmallCount
The minimum count for a covariate to appear in the data to be kept.

studyStartDate
A character object specifying the minimum date where data is used. Date format is 'yyyymmdd'.

studyEndDate
A character object specifying the maximum date where data is used. Date format is 'yyyymmdd'.

maxCasesPerOutcome
If there are more than this number of cases for a single outcome cases will be sampled to this size. maxCasesPerOutcome = 0 indicates no maximum size.

exposureIds
A list of identifiers to extract from the exposure table. If exposureTable = DRUG_ERA, exposureIds should be CONCEPT_ID. If exposureTable = "drug_era", exposureIds is used to select the drug_concept_id. If no exposure IDs are provided, all drugs or cohorts in the exposureTable are included as exposures.

customCovariateIds
A list of cohort definition IDs identifying the records in the customCovariateTable to use for building custom covariates.

Details

Create an object defining the parameter values.

[Package SelfControlledCaseSeries version 4.2.0 Index]