createSccsDiagnosticThresholds {SelfControlledCaseSeries}	R Documentation

Create SCCS diagnostics thresholds

Description

Threshold used when calling exportToCsv() to determine if we pass or fail diagnostics.

Usage

createSccsDiagnosticThresholds(
  mdrrThreshold = 10,
  easeThreshold = 0.25,
  timeTrendPThreshold = 0.05,
  preExposurePThreshold = 0.05
)

Arguments

mdrrThreshold
What is the maximum allowed minimum detectable relative risk (MDRR)?

easeThreshold
What is the maximum allowed expected absolute systematic error (EASE).

timeTrendPThreshold
What family-wise p-value threshold (alpha) will be used to determine temporal instability?

preExposurePThreshold
What p-value threshold (alpha) will be used to determine whether the rate of the outcome was higher just before exposure initiation?

Value

An object of type SccsDiagnosticThresholds.

[Package SelfControlledCaseSeries version 4.2.0 Index]