createScriIntervalData {SelfControlledCaseSeries}	R Documentation

Create Self-Controlled Risk Interval (SCRI) era data

Description

Create Self-Controlled Risk Interval (SCRI) era data

Usage

createScriIntervalData(
  studyPopulation,
  sccsData,
  eraCovariateSettings,
  controlIntervalSettings
)

Arguments

studyPopulation
An object created using the createStudyPopulation() function.

sccsData
An object of type SccsData as created using the getDbSccsData function.

eraCovariateSettings
Either an object of type EraCovariateSettings as created using the createEraCovariateSettings() function, or a list of such objects.

controlIntervalSettings
An object of type ControlIntervalSettings as created using the createControlIntervalSettings() function.

Details

This function creates interval data according to the elf-Controlled Risk Interval (SCRI) design. Unlike the generic SCCS design, where all patient time is used to establish a background rate, in the SCRI design a specific control interval (relative to the exposure) needs to be defined. The final model will only include time that is either part of the risk interval (defined using the eraCovariateSettings argument, or the control interval (defined using controlIntervalSettings).

Value

An object of type SccsIntervalData.

References

Greene SK, Kulldorff M, Lewis EM, Li R, Yin R, Weintraub ES, Fireman BH, Lieu TA, Nordin JD, Glanz JM, Baxter R, Jacobsen SJ, Broder KR, Lee GM. Near real-time surveillance for influenza vaccine safety: proof-of-concept in the Vaccine Safety Datalink Project. Am J Epidemiol. 2010 Jan 15;171(2):177-88. doi: 10.1093/aje/kwp345.

[Package SelfControlledCaseSeries version 4.2.0 Index]