Fit the SCCS model

Fit the SCCS model

Usage

fitSccsModel(sccsIntervalData, fitSccsModelArgs)

Arguments

sccsIntervalData

An object of type SccsIntervalData as created using the createSccsIntervalData function.

fitSccsModelArgs

An object of type FitSccsModelArgs as created by the createFitSccsModelArgs() function.

Value

An object of type SccsModel. Generic functions print, coef, and confint are available.

Details

Fits the SCCS model as a conditional Poisson regression. When allowed, coefficients for some or all covariates can be regularized.

Likelihood profiling is only done for variables for which profileLikelihood is set to TRUE when calling createEraCovariateSettings(). Either specify the profileGrid for a completely user- defined grid, or profileBounds for an adaptive grid. Both should be defined on the log IRR scale. When both profileGrid and profileGrid are NULL likelihood profiling is disabled.

References

Suchard, M.A., Simpson, S.E., Zorych, I., Ryan, P., and Madigan, D. (2013). Massive parallelization of serial inference algorithms for complex generalized linear models. ACM Transactions on Modeling and Computer Simulation 23, 10