fitSccsModel {SelfControlledCaseSeries}	R Documentation

Fit the SCCS model

Description

Fit the SCCS model

Usage

fitSccsModel(
  sccsIntervalData,
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", selectorType = "byPid", startingVariance =
    0.1, seed = 1, resetCoefficients = TRUE, noiseLevel = "quiet"),
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10))
)

Arguments

sccsIntervalData
An object of type SccsIntervalData as created using the createSccsIntervalData function.

prior
The prior used to fit the model. See Cyclops::createPrior for details.

control
The control object used to control the cross-validation used to determine the hyperparameters of the prior (if applicable). See Cyclops::createControl for details.

profileGrid
A one-dimensional grid of points on the log(relative risk) scale where the likelihood for coefficient of variables is sampled. See details.

profileBounds
The bounds (on the log relative risk scale) for the adaptive sampling of the likelihood function.

Details

Fits the SCCS model as a conditional Poisson regression. When allowed, coefficients for some or all covariates can be regularized.

Likelihood profiling is only done for variables for which profileLikelihood is set to TRUE when calling createEraCovariateSettings(). Either specify the profileGrid for a completely user- defined grid, or profileBounds for an adaptive grid. Both should be defined on the log IRR scale. When both profileGrid and profileGrid are NULL likelihood profiling is disabled.

Value

An object of type SccsModel. Generic functions print, coef, and confint are available.

References

Suchard, M.A., Simpson, S.E., Zorych, I., Ryan, P., and Madigan, D. (2013). Massive parallelization of serial inference algorithms for complex generalized linear models. ACM Transactions on Modeling and Computer Simulation 23, 10

[Package SelfControlledCaseSeries version 4.2.0 Index]