Run a list of analyses

Run a list of analyses

runSccsAnalyses(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  customCovariateDatabaseSchema = cdmDatabaseSchema,
  customCovariateTable = "cohort",
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "cohort",
  cdmVersion = 5,
  outputFolder = "./SccsOutput",
  sccsAnalysisList,
  exposureOutcomeList,
  combineDataFetchAcrossOutcomes = TRUE,
  getDbSccsDataThreads = 1,
  createStudyPopulationThreads = 1,
  createSccsIntervalDataThreads = 1,
  fitSccsModelThreads = 1,
  cvThreads = 1
)

Arguments

connectionDetails
An R object of type ConnectionDetails created using the function DatabaseConnector::createConnectionDetails().

cdmDatabaseSchema
The name of the database schema that contains the OMOP CDM instance. Requires read permissions to this database. On SQL Server, this should specify both the database and the schema, so for example 'cdm_instance.dbo'.

oracleTempSchema
DEPRECATED: use tempEmulationSchema instead.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

exposureDatabaseSchema
The name of the database schema that is the location where the exposure data used to define the exposure cohorts is available. If exposureTable = "DRUG_ERA", exposureDatabaseSchema is not used but assumed to be cdmDatabaseSchema. Requires read permissions to this database.

exposureTable
The table name that contains the exposure cohorts. If exposureTable <> "DRUG_ERA", then expectation is exposureTable has format of COHORT table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

outcomeDatabaseSchema
The name of the database schema that is the location where the data used to define the outcome cohorts is available. If outcomeTable = "CONDITION_ERA", outcomeDatabaseSchema is not used. Requires read permissions to this database.

outcomeTable
The table name that contains the outcome cohorts. If outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA, then expectation is outcomeTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

customCovariateDatabaseSchema
The name of the database schema that is the location where the custom covariate data is available.

customCovariateTable
Name of the table holding the custom covariates. This table should have the same structure as the cohort table.

nestingCohortDatabaseSchema
The name of the database schema that is the location where the nesting cohort is defined.

nestingCohortTable
Name of the table holding the nesting cohort. This table should have the same structure as the cohort table.

cdmVersion
Define the OMOP CDM version used: currently support "4" and "5".

outputFolder
Name of the folder where all the outputs will written to.

sccsAnalysisList
A list of objects of sccsAnalysis as created using the createSccsAnalysis() function.

exposureOutcomeList
A list of objects of type exposureOutcome as created using the createExposureOutcome() function.

combineDataFetchAcrossOutcomes
Should fetching data from the database be done one outcome at a time, or for all outcomes in one fetch? Combining fetches will be more efficient if there is large overlap in the subjects that have the different outcomes.

getDbSccsDataThreads
The number of parallel threads to use for building the SccsData objects.

createStudyPopulationThreads
The number of parallel threads to use for building the studyPopulation objects.

createSccsIntervalDataThreads
The number of parallel threads to use for building the SccsIntervalData objects.

fitSccsModelThreads
The number of parallel threads to use for fitting the models.

cvThreads
The number of parallel threads to use for the cross- validation when estimating the hyperparameter for the outcome model. Note that the total number of CV threads at one time could be fitSccsModelThreads * cvThreads.

Value

A tibble describing for each exposure-outcome-analysisId combination where the intermediary and outcome model files can be found, relative to the outputFolder.

Details

Run a list of analyses for the drug-comparator-outcomes of interest. This function will run all specified analyses against all hypotheses of interest, meaning that the total number of outcome models is length(cmAnalysisList) * length(drugComparatorOutcomesList) (if all analyses specify an outcome model should be fitted). When you provide several analyses it will determine whether any of the analyses have anything in common, and will take advantage of this fact. For example, if we specify several analyses that only differ in the way the outcome model is fitted, then this function will extract the data and fit the propensity model only once, and re-use this in all the analysis.